fix: hooksをutilsに移行
